Memoize signal, simplify example
The problem with the example was with react hook dependenies in that
the `trigger` function was stale, and so it was triggering an obsolete
signal. At least that was my hunch, so I refactored it to simplify and
memoize the stream of values, and it started working.

This eliminates the abstract autocomplete hook and just inlines the
example into place. I figured this was warranted since this is an
example and the most important thing is to showcase the autocomplete
logic as directly as possible.

I extracted the job of setting up and teardig down a task into a
`useTask()` hook so that it is secondary